+++ Character/CharacterStats.cpp
@@ -57,7 +57,7 @@ double CharacterStats::get_hit_chance() const {
 
 double CharacterStats::get_crit_chance() const {
     const double equip_effect = base_stats->get_crit_chance()  + equipment->get_stats()->get_crit_chance();
-    const auto crit_from_agi = double(double(get_agility()) / pchar->get_agi_needed_for_one_percent_phys_crit());
+    const auto crit_from_agi = double(get_agility()) / pchar->get_agi_needed_for_one_percent_phys_crit();
 
     return equip_effect + crit_from_agi / 100;
 }
@@ -68,7 +68,7 @@ double CharacterStats::get_spell_hit_chance() const {
 
 double CharacterStats::get_spell_crit_chance() const {
     const double equip_effect = base_stats->get_spell_crit_chance()  + equipment->get_stats()->get_spell_crit_chance();
-    const auto crit_from_int = double(double(get_intellect()) / pchar->get_int_needed_for_one_percent_spell_crit());
+    const auto crit_from_int = double(get_intellect()) / pchar->get_int_needed_for_one_percent_spell_crit();
 
     return equip_effect + crit_from_int / 100;
 }
